[#1848] Improve Members, NembershipTasks tests
1. Member-related tests moved to a separate test file
2. Added layout tests for MemberForm and MembershipTaskForm
